CI: Move FCB tests to file
